requirements.txt
socketIO-client==0.5.7.4
pyyaml
